Moved missed functions from GMP class to TaskMixin
